Version 2.4: So easy to use.

+++ lib/src/main/java/com/github/chenyoca/validation/supports/StringLazyLoader.java
@@ -10,8 +10,8 @@ import com.github.chenyoca.validation.LazyLoader;
 public abstract class StringLazyLoader implements LazyLoader {
 
     @Override
-    final public int[] intValues() {
-        return new int[0];
+    final public long[] intValues() {
+        return new long[0];
     }
 
     @Override
